+++ frontend/src/index.ts
@@ -1,4 +1,4 @@
-import createCheckBox from './components/checkbox';
+import createCheckBox from './components/checkbox/checkbox';
 
 export function sum(a: number, b: number): HTMLElement {
   return document.createElement('div');
